+++ tests/unit/test_responses.py
@@ -221,8 +221,8 @@ def test_responses_json(name: str, response_class: _ResponseBase):
 )
 def test_responses_json_orjson_encoder(name: str, response_class: _ResponseBase):
     """Tests that json() works correctly with orjson encoder."""
+    orjson = pytest.importorskip("orjson")
     import orjson
-
     response = read_response_file(name)
     mocked = mock_response(response)
     response_data = response_class.read_response(mocked)